requirements.txt
asgiref==3.2.10
coverage==5.3
Django==3.1.1
pytz==2020.1
sqlparse==0.3.1

django-crispy-forms~=1.13.0